Refactor code to use API to expand TDEPS
The first implementation used a methodology very similar to the
install command, where each package was downloaded, and the TDEPS were
gathered from that downloaded package. That ended up being overly
complex, replicated code in a way that was awkward to refactor and
blocked any future work around parallelism.

This is a different take that simplifies the whole process substantially.

TDEPS and resolved idents are gathered from the builder API, and a
download manifest is generated. That manifest is then downloaded as a
separate phase.

Signed-off-by: Mark Anderson <mark@chef.io>